incluir arquivo uid.txt

+++ js/app.js
@@ -1,3 +1,4 @@
+dadosUser = document.querySelector("#dadosUser")
 
 // Função para enviar arquivo para o Firebase Storage
 function uploadFile() {
@@ -90,38 +91,46 @@ function displayFiles() {
         result.items.forEach(item => {
           item.getDownloadURL().then(url => {
             item.getMetadata().then(metadata => {
-              var fileContainer = document.createElement('div');
-              fileContainer.classList.add('file-container'); // Adicionar a classe
 
-              if (item.name.toLowerCase().endsWith('.jpg') || item.name.toLowerCase().endsWith('.jpeg') || item.name.toLowerCase().endsWith('.png')) {
-                // Se for uma imagem, exibir miniatura
-                var img = document.createElement('img');
-                img.src = url; // Usar a variável url diretamente
-                img.alt = metadata.name; // Adicione o nome do arquivo como texto alternativo
-                img.classList.add('file-image'); // Adicionar a classe
-                fileContainer.appendChild(img);
-              } else {
-                // Se for outro tipo de arquivo, exibir um link direto para o download
-                var link = document.createElement('a');
-                link.href = url; // Usar a variável url diretamente
-                link.innerHTML = metadata.name; // Use metadata.name para exibir o nome do arquivo
-                fileContainer.appendChild(link);
+              let arquivoconfig = user.uid + ".txt"
+              if (metadata.name != arquivoconfig) {
+
+                var fileContainer = document.createElement('div');
+                fileContainer.classList.add('file-container'); // Adicionar a classe
+
+                if (item.name.toLowerCase().endsWith('.jpg') || item.name.toLowerCase().endsWith('.jpeg') || item.name.toLowerCase().endsWith('.png')) {
+                  // Se for uma imagem, exibir miniatura
+                  var img = document.createElement('img');
+                  img.src = url; // Usar a variável url diretamente
+                  img.alt = metadata.name; // Adicione o nome do arquivo como texto alternativo
+                  img.classList.add('file-image'); // Adicionar a classe
+                  fileContainer.appendChild(img);
+                } else {
+                  // Se for outro tipo de arquivo, exibir um link direto para o download
+                    var link = document.createElement('a');
+                    link.href = url; // Usar a variável url diretamente
+                    link.innerHTML = metadata.name; // Use metadata.name para exibir o nome do arquivo
+                    fileContainer.appendChild(link);
+                }
+
+                // Adicione o botão de remoção
+                var removeButton = document.createElement('button');
+                removeButton.textContent = 'Remover';
+                removeButton.addEventListener('click', function() {
+                  removeFile(user.uid, item.name);
+                });
+                fileContainer.appendChild(removeButton);
+
+                // Adicione o nome do arquivo abaixo da miniatura ou link
+                var fileNameElement = document.createElement('p');
+                fileNameElement.textContent = metadata.name;
+                fileContainer.appendChild(fileNameElement);
+
+                fileList.appendChild(fileContainer);
               }
-
-              // Adicione o botão de remoção
-              var removeButton = document.createElement('button');
-              removeButton.textContent = 'Remover';
-              removeButton.addEventListener('click', function() {
-                removeFile(user.uid, item.name);
-              });
-              fileContainer.appendChild(removeButton);
-
-              // Adicione o nome do arquivo abaixo da miniatura ou link
-              var fileNameElement = document.createElement('p');
-              fileNameElement.textContent = metadata.name;
-              fileContainer.appendChild(fileNameElement);
-
-              fileList.appendChild(fileContainer);
+              else {
+                  dadosUser.innerHTML = arquivoconfig
+              }
             });
           }).catch(error => {
             console.error('Erro ao recuperar metadados:', error);
